Restructure directories and update routes/links
- Move images to an images folder
- Move `chapters.json` and `Chapters.vue` to a new `components/chapters`
folder
- Move `Workshops.vue` and `ExerciseLink.vue` to a new
`components/workshops`
folder
- Move `Header.vue` to `components` directory
- Move all images to new `images` directory and update images links in
other files
- Update `main.js` with new file structure where needed
