+++ src/components/Lesson.vue
@@ -2,10 +2,10 @@
   <div>
     <header class="bg-navy pa2 pa3-ns flex items-center justify-around">
       <a href='/#/' class="db-ns link flex-auto">
-        <img src="./home/ps_logo_horiz_white.svg" alt="ProtoSchool" style="height: 80px" class="ml3-ns"/>
+        <img src="../images/ps_logo_horiz_white.svg" alt="ProtoSchool" style="height: 80px" class="ml3-ns"/>
       </a>
       <div class="tr dn db-ns">
-        <img src="./home/ipfs-illustrations-how-3.svg" alt="" style="height: 50px">
+        <img src="../images/ipfs-illustrations-how-3.svg" alt="" style="height: 50px">
       </div>
     </header>
 
@@ -14,7 +14,7 @@
         <h1 class="f3 measure-wide">{{lessonTitle}}</h1>
         <div class="lh-solid v-mid">
           <span class="green v-mid"><span class="b">{{workshopShortname}}</span> | Lesson {{lessonNumber}} of {{lessonsInWorkshop}}</span>
-          <span class="pl1"><img v-if="lessonPassed" src="./home/complete.svg" alt="complete" style="height: 1.2rem;" class="v-mid"/></span>
+          <span class="pl1"><img v-if="lessonPassed" src="../images/complete.svg" alt="complete" style="height: 1.2rem;" class="v-mid"/></span>
         </div>
         <div class="lesson-text lh-copy measure-wide" v-html="parsedText"></div>
       </section>
@@ -28,9 +28,9 @@
       <div class="flex-none">
         <h2 class="mt0 mb2 green fw4 fill-current">
           <span style='vertical-align:-1px'>
-            <img v-if="lessonPassed" src="./home/complete.svg" alt="complete" style="height: 1rem;"/>
-            <img v-else-if="cachedCode" src="./home/in-progress.svg" alt="complete" style="height: 1rem;"/>
-            <img v-else src="./home/not-started.svg" alt="not yet started" style="height: 1rem;"/>
+            <img v-if="lessonPassed" src="../images/complete.svg" alt="complete" style="height: 1rem;"/>
+            <img v-else-if="cachedCode" src="../images/in-progress.svg" alt="complete" style="height: 1rem;"/>
+            <img v-else src="../images/not-started.svg" alt="not yet started" style="height: 1rem;"/>
           </span>
           <span class="green ttu f6 pl2 pr1 fw7 v-mid">
             <span v-if="lessonPassed">You did it!</span>
