+++ src/components/home/Home.vue
@@ -12,7 +12,7 @@
 </template>
 
 <script>
-import Header from './Header'
+import Header from '../Header'
 
 export default {
   name: 'home',
