+++ src/components/workshops/ExerciseLink.vue
@@ -3,9 +3,9 @@
     <div class="flex">
       <div class="green ttu f6" style="min-width: 72px">Lesson {{index}}</div>
       <div class="pr2">
-        <img v-if="lessonPassed('passed' + to)" src="./complete.svg" alt="complete" style="height: 1rem;"/>
-        <img v-else-if="lessonCached('cached' + to)" src="./in-progress.svg" alt="complete" style="height: 1rem;"/>
-        <img v-else src="./not-started.svg" alt="not yet started" style="height: 1rem;"/>
+        <img v-if="lessonPassed('passed' + to)" src="../../images/complete.svg" alt="complete" style="height: 1rem;"/>
+        <img v-else-if="lessonCached('cached' + to)" src="../../images/in-progress.svg" alt="complete" style="height: 1rem;"/>
+        <img v-else src="../../images/not-started.svg" alt="not yet started" style="height: 1rem;"/>
       </div>
       <div class="navy fw5 mw6">{{name}}</div>
     </div>
