+++ src/main.js
@@ -6,8 +6,8 @@ import Vue from 'vue'
 import VueRouter from 'vue-router'
 import App from './App.vue'
 import Home from './components/home/Home.vue'
-import Chapters from './components/home/Chapters.vue'
-import Workshops from './components/home/Workshops.vue'
+import Chapters from './components/chapters/Chapters.vue'
+import Workshops from './components/workshops/Workshops.vue'
 import LessonBasics01 from './lessons/Basics/01.vue'
 import LessonBasics02 from './lessons/Basics/02.vue'
 import LessonBasics03 from './lessons/Basics/03.vue'
